[FIX] swap parameter
