+++ spp_base_common/__manifest__.py
@@ -33,7 +33,7 @@
     "assets": {
         "web.assets_backend": [
             "spp_base_common/static/src/scss/navbar.scss",
-            "spp_base_common/static/src/js/list_controller.js",
+            "spp_base_common/static/src/legacy/js/views/list/list_controller_fix.js",
         ],
         "web._assets_primary_variables": [
             "spp_base_common/static/src/scss/colors.scss",
